Add files via upload
The "ML" pictures are pictures that include a logo (Met Logo). ZL is "Zonder Logo" and thus exclude the logo. The files for data procession are added too. Aankoopgedrag (Responses) is the original, and from the "kopie" I added a column with the total willingness (of the 8 products). 50+ is the one used for the people-above-50-analysis. "...to see buying intentions when voted yes" is the file used for the analysis of the buying intentions of the people who said they normally buy ecologically labelled products.